T496: small code cleaning
Rewrote some small things, such as setting the default value of 'default_value_from_file'
to None instead of 0, and factored out the getattr to clean code a little
into smaller lines.
I also removed the except AttributeError since it's not required considering
that we don't use getattr anymore with a variable key.